+++ packages/preprocessor/CHANGELOG.md
@@ -1,6 +1,7 @@
 ### HEAD
 
 * Refactor to support Node 4 upwards
+* Update `postcss-easy-import` to `2.0.0`
 * Update `fs-extra` to `2.1.2`
 * Update `autoprefixer` to `6.7.7`
 * Update `postcss-apply` to `0.6.1`